2.1.0: Merged bug fixes from 2.0.2.
git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/2.1.0@1798 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ modules/ffmpeg/producer/audio/audio_decoder.h
@@ -49,9 +49,9 @@ public:
 	void push(const std::shared_ptr<AVPacket>& packet);
 	std::shared_ptr<core::audio_buffer> poll();
 
-	int64_t nb_frames() const;
+	uint32_t nb_frames() const;
 	
-	size_t file_frame_number() const;
+	uint32_t file_frame_number() const;
 private:
 	struct implementation;
 	safe_ptr<implementation> impl_;
